COMP: Fix sign-compare warning in vtkSlicerVolumesLogic
This commit fixes are a warning introduced by r23934 (BUG: avoid false
positives on geometry check)

// -----------
/path/to/Slicer/Modules/Loadable/Volumes/Logic/vtkSlicerVolumesLogic.cxx:96:41: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for (int i = 0; i < fractional.length(); ++i)
                                         ^
/path/to/Slicer/Modules/Loadable/Volumes/Logic/vtkSlicerVolumesLogic.cxx:114:36: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     if (i == fractional.length() - 1)
// -----------                                    ^

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=23947
git-svn-id: http://svn.slicer.org/Slicer4/trunk@23947 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/Loadable/Volumes/Logic/vtkSlicerVolumesLogic.cxx
@@ -93,7 +93,7 @@ int significantDecimals(double value, int defaultDecimals = -1)
   int previousRepeat=0;
   bool only0s = true;
   bool isUnit = value > -1. && value < 1.;
-  for (int i = 0; i < fractional.length(); ++i)
+  for (size_t i = 0; i < fractional.length(); ++i)
     {
     char digit = fractional.at(i);
     if (digit != '0')
